Clean up code
git-svn-id: svn+ssh://mail.hekkelman.net/home/maarten/svnroot/libzeep/trunk@327 31ceaa29-a5dd-4727-bcb8-6a3098495666

+++ src/request_parser.cpp
@@ -40,7 +40,7 @@ void request_parser::reset()
 }
 
 boost::tribool request_parser::parse(request& req,
-	const char* text, unsigned int length)
+	const char* text, size_t length)
 {
 	if (m_parser == NULL)
 	{
